requirements.txt
PyMuPDF==1.19.6  
Pillow==9.1.0  
pytesseract==0.3.8
spacy==3.2.1

